@@ -13,59 +13,65 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-@CrossOrigin(origins = { "http://localhost:3000", "http://localhost:4200" })
+@CrossOrigin(origins = {"http://localhost:3000", "http://localhost:4200"})
 @RestController
+@RequestMapping("/instructors/{username}/courses")
 public class CourseResource {
 
-	@Autowired
-	private CoursesHardcodedService courseManagementService;
+    @Autowired
+    private CoursesHardcodedService courseManagementService;
 
-	@GetMapping("/instructors/{username}/courses")
-	public List<Course> getAllCourses(@PathVariable String username) {
-		return courseManagementService.findAll();
-	}
+    @GetMapping()
+    public List<Course> getAllCourses(@PathVariable String username) {
+        return courseManagementService.findAll();
+    }
 
-	@GetMapping("/instructors/{username}/courses/{id}")
-	public Course getCourse(@PathVariable String username, @PathVariable long id) {
-		return courseManagementService.findById(id);
-	}
+    @GetMapping("/{id}")
+    public Course getCourse(@PathVariable String username,
+                            @PathVariable long id) {
+        return courseManagementService.findById(id);
+    }
 
-	@DeleteMapping("/instructors/{username}/courses/{id}")
-	public ResponseEntity<Void> deleteCourse(@PathVariable String username, @PathVariable long id) {
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteCourse(@PathVariable String username,
+                                             @PathVariable long id) {
 
-		Course course = courseManagementService.deleteById(id);
+        Course course = courseManagementService.deleteById(id);
 
-		if (course != null) {
-			return ResponseEntity.noContent().build();
-		}
+        if (course != null) {
+            return ResponseEntity.noContent().build();
+        }
 
-		return ResponseEntity.notFound().build();
-	}
+        return ResponseEntity.notFound().build();
+    }
 
-	@PutMapping("/instructors/{username}/courses/{id}")
-	public ResponseEntity<Course> updateCourse(@PathVariable String username, @PathVariable long id,
-			@RequestBody Course course) {
+    @PutMapping("/{id}")
+    public ResponseEntity<Course> updateCourse(@PathVariable String username,
+                                               @PathVariable long id,
+                                               @RequestBody Course course) {
 
-		Course courseUpdated = courseManagementService.save(course);
+        Course courseUpdated = courseManagementService.save(course);
 
-		return new ResponseEntity<Course>(course, HttpStatus.OK);
-	}
+        return new ResponseEntity<Course>(course, HttpStatus.OK);
+    }
 
-	@PostMapping("/instructors/{username}/courses")
-	public ResponseEntity<Void> createCourse(@PathVariable String username, @RequestBody Course course) {
+    @PostMapping()
+    public ResponseEntity<Void> createCourse(@PathVariable String username, @RequestBody Course course) {
+        Course createdCourse = courseManagementService.save(course);
 
-		Course createdCourse = courseManagementService.save(course);
-
-		// Location
-		// Get current resource url
-		/// {id}
-		URI uri = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}").buildAndExpand(createdCourse.getId())
-				.toUri();
-
-		return ResponseEntity.created(uri).build();
-	}
+        // Location
+        // Get current resource url
+        /// {id}
+        URI uri = ServletUriComponentsBuilder.fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(createdCourse.getId())
+                .toUri();
+        System.out.println(courseManagementService.findById(createdCourse.getId()));
+        return ResponseEntity.created(uri).build();
+    }
 
 }
