@@ -8,20 +8,19 @@ import org.springframework.stereotype.Service;
 @Service
 public class CoursesHardcodedService {
 
-	private static List<Course> courses = new ArrayList<>();
-	private static long idCounter = 0;
+    private static final List<Course> courses = new ArrayList<>();
+    private static long idCounter = 0;
 
-	static {
-		courses.add(new Course(++idCounter, "in28minutes", "Learn Full stack with Spring Boot and Angular"));
-		courses.add(new Course(++idCounter, "in28minutes", "Learn Full stack with Spring Boot and React"));
-		courses.add(new Course(++idCounter, "in28minutes", "Master Microservices with Spring Boot and Spring Cloud"));
-		courses.add(new Course(++idCounter, "in28minutes",
-				"Deploy Spring Boot Microservices to Cloud with Docker and Kubernetes"));
-	}
+    static {
+        courses.add(new Course(++idCounter, "in28minutes", "Learn Full stack with Spring Boot and Angular"));
+        courses.add(new Course(++idCounter, "in28minutes", "Learn Full stack with Spring Boot and React"));
+        courses.add(new Course(++idCounter, "in28minutes", "Master Microservices with Spring Boot and Spring Cloud"));
+        courses.add(new Course(++idCounter, "in28minutes", "Deploy Spring Boot Microservices to Cloud with Docker and Kubernetes"));
+    }
 
-	public List<Course> findAll() {
-		return courses;
-	}
+    public List<Course> findAll() {
+        return courses;
+    }
 
 	public Course save(Course course) {
 		if (course.getId() == -1 || course.getId() == 0) {
@@ -34,26 +33,23 @@ public class CoursesHardcodedService {
 		return course;
 	}
 
-	public Course deleteById(long id) {
-		Course course = findById(id);
+    public Course deleteById(long id) {
+        Course course = findById(id);
 
-		if (course == null)
-			return null;
+        if (course == null)
+            return null;
 
-		if (courses.remove(course)) {
-			return course;
-		}
+        if (courses.remove(course)) {
+            return course;
+        }
 
-		return null;
-	}
+        return null;
+    }
 
-	public Course findById(long id) {
-		for (Course course : courses) {
-			if (course.getId() == id) {
-				return course;
-			}
-		}
-
-		return null;
-	}
+    public Course findById(long id) {
+        return courses.stream()
+                .filter(course -> course.getId().equals(id))
+                .findFirst()
+                .orElse(null);
+    }
 }
